Add files via upload
Problem Statement:
A digital media company (similar to Voot, Hotstar, Netflix, etc.) had launched a show. Initially, the show got a good response, but then witnessed a decline in viewership. The company wants to figure out Approach:
We are concerned about determining the driver variable for show viewership. This is the case of prediction rather than projection where we are more interested in predicting the key driver variables and their impact rather than forecasting the results.

First we will list down the potential reasons for the decline in viewership.

The potential reasons could be:

Decline in the number of people coming to the platform
Fewer people watching the video
A Decrease in marketing spend?
Competitive shows, e.g. cricket/ IPL
Special holidays
Twist in the story at went wrong.

Data
We have been given data for the period of 1 March 2017 to 19 May 2017.
With Columns as
Views_show : Number of times the show was viewed
Visitors : Number of visitors who browsed the platform, but not necessarily watched a video.
Views_platform : Number of times a video was viewed on the platform
Ad_impression : Proxy for marketing budget. Represents number of impressions generated by ads
Cricket_match_india: If a cricket match was being played. 1 indicates match on a given day, 0 indicates there wasn't
Character_A : Describes presence of Character A. 1 indicates character A was in the episode, 0 indicates she/he wasn't.